Start implementing a subset of Apache Velocity with the intent of migrating AutoValue from Velocity to this reimplementation. The Velocity Template Language is a good way of expressing code generation, especially when using any of the many editors that support it; but Apache Velocity has caused many headaches, especially because of its pervasive use of reflection. In addition to the many workarounds visible at http://[] we have been unable to maven-shade Velocity, which causes problems when other libraries are present that use a different Velocity version. (See https://github.com/google/auto/issues/210.)
I have a complete prototype implementation at https://github.com/eamonnmcmanus/auto/tree/escapevelocity but I am proposing to break it down into somewhat more manageable chunks for code review. This chunk defines the basic API for creating and using templates, but only supports a small subset of the language. The next steps are to add support for expressions and for directives (#if, #set, etc). Support for macros should probably be in a separate step, because of the non-obvious nature of the call-by-name semantics they use. Then the final step will be to switch AutoValue to using the reimplementation, and ditch Velocity.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=94517530